refactor: getCategories query
