+++ src/apis/types/dto/category.dto.ts
@@ -1,3 +1,3 @@
-import { Category } from '@/apis/fetchers/category/get-categories'
+import { Category } from '@/apis/fetchers/category/get-categories/fetcher'
 
 export interface CategoryDTO extends Category {}
