+++ src/app/(routes)/(tabs)/repository/components/category-item.tsx
@@ -8,7 +8,7 @@ import CategoryTag from './category-tag'
 import Link from 'next/link'
 import Image from 'next/image'
 import icons from '@/constants/icons'
-import { Category } from '@/apis/fetchers/category/get-categories'
+import { Category } from '@/apis/fetchers/category/get-categories/fetcher'
 import { useSortable } from '@dnd-kit/sortable'
 import { CSS } from '@dnd-kit/utilities'
 import { cn } from '@/lib/utils'
